backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Config
python-dotenv==1.0.1
pydantic-settings==2.5.2
pydantic[email]==2.9.2

# HTTP client (llamadas a APIs externas)
httpx==0.27.2

# AI
anthropic==0.34.2

# Task queue (barridos automáticos)
celery==5.4.0
redis==5.0.8

# Utilities
python-slugify==8.0.4
